Se carga al repositorio la entrega del proyecto final

+++ src/routes/api/product.routes.js
@@ -1,6 +1,8 @@
+import passportCall from "../../utils/passportCall.js"
 import { Router } from "express";
 import ProductManager from "../../managers/product.manager.js";
 import uploader from "../../utils/uploader.js";
+import ProductDto from "../../dao/dto/product.dto.js";
 
 import {
     ERROR_INVALID_ID,
@@ -21,7 +23,8 @@ const handleError = (res, message) => {
 router.get("/", async (req, res) => {
     try {
         const products = await productManager.getAll(req.query);
-        res.status(200).json({ status: true, payload: products });
+        const productsDto = products.docs.map(product => new ProductDto(product))
+        res.status(200).json({ status: true, payload: productsDto });
     } catch (error) {
         handleError(res, error.message);
     }
@@ -31,41 +34,63 @@ router.get("/", async (req, res) => {
 router.get("/:id", async (req, res) => {
     try {
         const product = await productManager.getOneById(req.params.id);
-        res.status(200).json({ status: true, payload: product });
+        const productDto = new ProductDto(product)
+        res.status(200).json({ status: true, payload: productDto });
     } catch (error) {
         handleError(res, error.message);
     }
 });
 
 // Ruta para crear un nuevo producto, permite la subida de archivos
-router.post("/", uploader.single("file"), async (req, res) => {
-    try {
-        const { file } = req;
-        const product = await productManager.insertOne(req.body, file?.filename);
-        res.status(201).json({ status: true, payload: product });
-    } catch (error) {
-        handleError(res, error.message);
+router.post("/", uploader.single("file"), passportCall("current"), async (req, res) => {
+    if (req.user && req.user.role == "admin") {
+        try {
+            const { file } = req;
+            const product = await productManager.insertOne(req.body, file?.filename);
+            const productDto = new ProductDto(product);
+            res.status(201).json({ status: true, payload: productDto });
+        } catch (error) {
+            handleError(res, error.message);
+        }
+    } else {
+        res.status(401).json({
+            message: "Unauthorized access",
+        });
     }
 });
 
 // Ruta para actualizar un producto por su ID, permite la subida de archivos
 router.put("/:id", uploader.single("file"), async (req, res) => {
-    try {
-        const { file } = req;
-        const product = await productManager.updateOneById(req.params.id, req.body, file?.filename);
-        res.status(200).json({ status: true, payload: product });
-    } catch (error) {
-        handleError(res, error.message);
+    if (req.user && req.user.role == "admin") {
+        try {
+            const { file } = req;
+            const product = await productManager.updateOneById(req.params.id, req.body, file?.filename);
+            const productDto = new ProductDto(product)
+            res.status(200).json({ status: true, payload: productDto });
+        } catch (error) {
+            handleError(res, error.message);
+        }
+    } else {
+        res.status(401).json({
+            message: "Unauthorized access",
+        });
     }
 });
 
 // Ruta para eliminar un producto por su ID
 router.delete("/:id", async (req, res) => {
-    try {
-        const product = await productManager.deleteOneById(req.params.id);
-        res.status(200).json({ status: true, payload: product });
-    } catch (error) {
-        handleError(res, error.message);
+    if (req.user && req.user.role == "admin") {
+        try {
+            const product = await productManager.deleteOneById(req.params.id);
+            const productDto = new ProductDto(product)
+            res.status(200).json({ status: true, payload: productDto });
+        } catch (error) {
+            handleError(res, error.message);
+        }
+    } else {
+        res.status(401).json({
+            message: "Unauthorized access",
+        });
     }
 });
